+++ src/Simulation/Simulators/QuantumSimulator/ApplyUncontrolledRy.cs
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
+using System;
 using Microsoft.Quantum.Simulation.Core;
 using Microsoft.Quantum.Intrinsic.Interfaces;
 
@@ -12,7 +13,7 @@ namespace Microsoft.Quantum.Simulation.Simulators
         {
             this.CheckQubit(target, nameof(target));
             CheckAngle(angle);
-            R(this.Id, Pauli.PauliY, angle, (uint)target.Id);
+            R(this.Id, Pauli.PauliY, angle, (IntPtr)target.Id);
         }
     }
 }
